feat(editor): add useYjsRoom and useEditorMount hooks; enhance useSessionAnalytics
- Introduced `useYjsRoom` for managing Yjs room lifecycle, including awareness tracking and role assignment.
- Added `useEditorMount` for handling Monaco editor mounting and Yjs binding, ensuring proper readOnly enforcement based on user roles.
- Updated `useSessionAnalytics` to accept an optional initial role parameter in `startSession` and added a `resetAnalytics` method for cleanup.
- Enhanced editor hooks barrel export to include new hooks.
