+++ app/hooks/editor/index.ts
@@ -2,6 +2,8 @@
  * Editor hooks barrel export
  */
 
+export { useYjsRoom } from './useYjsRoom'
+export { useEditorMount } from './useEditorMount'
 export { useSessionAnalytics } from './useSessionAnalytics'
 export { usePanelLayout } from './usePanelLayout'
 export { useFileOperations } from './useFileOperations'
